feat: add tests for sorting algorithms and refactor
- Add Jest tests for all sorting algorithms.
- Refactor sorting logic out of visualizer components and into pure functions in `src/lib/algorithms.ts`.
- Update visualizer components to use the new sorting functions.
